requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
grpcio==1.52.0
grpcio-tools==1.52.0
gTTS==2.3.1
idna==3.4
protobuf==4.22.1
requests==2.28.2
urllib3==1.26.15
